# lib/api-zod/src/generated/types/acta.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SG-SST Regis Colombia API
 * OpenAPI spec version: 0.1.0
 */

export interface Acta {
  id: string;
  empresa_id: string;
  numero_acta?: string | null;
  fecha_reunion?: string | null;
  estado: string;
  created_at: string;
}