# lib/api-zod/src/generated/types/examenProcesado.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SG-SST Regis Colombia API
 * OpenAPI spec version: 0.1.0
 */
import type { ExamenProcesadoRaw } from "./examenProcesadoRaw";
import type { ExamenProcesadoRecomendacionesItem } from "./examenProcesadoRecomendacionesItem";
import type { ExamenProcesadoRestriccionesItem } from "./examenProcesadoRestriccionesItem";

export interface ExamenProcesado {
  examen_id: string;
  concepto: string;
  restricciones: ExamenProcesadoRestriccionesItem[];
  recomendaciones: ExamenProcesadoRecomendacionesItem[];
  requiere_seguimiento: boolean;
  raw?: ExamenProcesadoRaw;
}