# lib/api-zod/src/generated/types/generarActaBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SG-SST Regis Colombia API
 * OpenAPI spec version: 0.1.0
 */
import type { GenerarActaBodyAsistentesItem } from "./generarActaBodyAsistentesItem";
import type { GenerarActaBodyTipoComite } from "./generarActaBodyTipoComite";

export interface GenerarActaBody {
  empresa_id: string;
  tipo_comite: GenerarActaBodyTipoComite;
  fecha: string;
  hora_inicio: string;
  hora_fin: string;
  lugar: string;
  asistentes: GenerarActaBodyAsistentesItem[];
  puntos: string[];
  transcripcion?: string | null;
}