# lib/api-zod/src/generated/types/matrizGenerada.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SG-SST Regis Colombia API
 * OpenAPI spec version: 0.1.0
 */
import type { MatrizGeneradaContenido } from "./matrizGeneradaContenido";
import type { MatrizGeneradaNivelesRiesgo } from "./matrizGeneradaNivelesRiesgo";

export interface MatrizGenerada {
  matriz_id: string;
  ciiu: string;
  actividad_economica: string;
  num_peligros: number;
  niveles_riesgo: MatrizGeneradaNivelesRiesgo;
  peligros_prioritarios: string[];
  contenido: MatrizGeneradaContenido;
}