# lib/api-zod/src/generated/types/pilaProcesada.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SG-SST Regis Colombia API
 * OpenAPI spec version: 0.1.0
 */
import type { PilaProcesadaNumAfiliados } from "./pilaProcesadaNumAfiliados";
import type { PilaProcesadaRaw } from "./pilaProcesadaRaw";

export interface PilaProcesada {
  registro_id: string;
  periodo: string;
  estado: string;
  num_afiliados: PilaProcesadaNumAfiliados;
  alertas: string[];
  raw?: PilaProcesadaRaw;
}